+++ app-skyfitnesspro/src/components/progress/progress.styled.js
